Moved arcane symbol data logic to service
